Some minor redux refactors
