+++ src/redux/reducers/index.ts
@@ -6,5 +6,4 @@ const reducers = combineReducers({
   user,
 })
 
-export type RootState = ReturnType<typeof reducers>
 export default reducers
