+++ src/redux/store.ts
@@ -12,4 +12,9 @@ function configureStore() {
   return store
 }
 
-export default configureStore()
+const store = configureStore()
+
+export type RootState = ReturnType<typeof store.getState>
+export type AppDispatch = typeof store.dispatch
+
+export default store
